feat(next-drupal): translatePath, getResourceByPath - throw error when backend error
Until now, any not-ok response from decoupled router was treated equally, as a 404 Not Found.
Behaviour is changed differentiating between 404 Not Found and server error.
The advantage is frontend holds to previous built page and doesn't update public page content
with a 404 due to backend issues.

BREAKING CHANGE:
Might break sites relying on response being the same in 404 than in 50x.

Fixes #687